--- src/components/common/ColonyCreationWizard/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ColonyCreationWizard';
