+++ src/components/common/CreateColonyWizard/CreateColonyCardRow.css
