+++ src/components/common/CreateColonyWizard/StepColonyName.css
