+++ src/components/common/CreateColonyWizard/StepColonyName.css.d.ts
