+++ src/components/common/CreateColonyWizard/StepColonyNameValidation.ts
