+++ src/components/common/CreateColonyWizard/StepConfirmAllInput.css
