+++ src/components/common/CreateColonyWizard/StepConfirmAllInput.css.d.ts
