+++ src/components/common/CreateColonyWizard/StepConfirmTransactions.css
