+++ src/components/common/CreateColonyWizard/StepConfirmTransactions.css.d.ts
