+++ src/components/common/CreateColonyWizard/StepCreateToken.css
