+++ src/components/common/CreateColonyWizard/StepSelectToken.css
