+++ src/components/common/CreateColonyWizard/StepTokenChoice.css
