+++ src/components/common/CreateColonyWizard/TokenSelector.css
