+++ src/components/common/CreateColonyWizard/TokenSelector.tsx
